Refactor SignInPage to use hooks + typescripts
